fix(tests): the UX fixture reuses the auto-created latest release
Closes #1261.

The fixture created its own release with is_latest=True, but adding the
SBOM and the document fires the signals that already create one, so the
second tripped the one-latest-release guard. 28 of 32 cases errored at
setup and the generator produced 4 of its 32 screenshots. get_or_create
on the reserved name, so it still works if the fixture is ever used
without artifacts.

The module is gated behind RUN_UX_REVIEW=1 and skipped in CI, which is
right for a screenshot tool but is why this rotted unseen. Moved
trust_center_product into e2e/fixtures.py, where every other factory
already lives, and added an ordinary non-gated test that builds it and
asserts exactly one latest release. That runs in CI, so the next
collision fails loudly instead of silently halving the output.

RUN_UX_REVIEW=1: 4 passed/28 errors -> 32 passed, 32 images.
